Separate the webpack configuration into a series of different files so that they’re more portable.
The resolution essentially goes:

    [base] -> [application] -> [environment]

Where `application` is the place to put your app-specific config. The
change also adds support for react, es6 (through babel) and better
listing through eslint.
