+++ .babelrc
@@ -1,0 +1,6 @@
+{
+  "presets": [
+    "babel-preset-es2015",
+    "babel-preset-react"
+  ]
+}
