+++ assets/webpack-application.config.js
@@ -1,0 +1,9 @@
+/**
+ * Application webpack configuration
+ *
+ * This will get merged with `webpack-base.config` and the environment specific
+ * config at build time using https://www.npmjs.com/package/webpack-merge
+ *
+ * @type {Object}
+ */
+module.exports = {}
