+++ assets/webpack-production.config.js
@@ -1,144 +1,29 @@
-var fs = require("fs");
-var glob = require("glob");
 var path = require("path");
 var webpack = require("webpack");
-var loadConfig = require("./config");
-
-/**
- * General configuration
- */
-var config  = loadConfig(path.join(__dirname, "config.yml"));
-var BASE      = path.join(__dirname, "..")
-var APPS_BASE = path.join(BASE, "/apps")
-var APPS      = glob.sync(APPS_BASE + "/*");
-var BUILD     = path.join(__dirname, "build");
-
-/**
- * Custom webpack plugins
- */
-var ExtractTextPlugin = require("extract-text-webpack-plugin");
-
-/**
- * PostCSS packages
- */
-var cssimport = require("postcss-import");
-var cssnext = require("postcss-cssnext");
-
-/**
- * createEntries
- *
- * Iterate through the `APPS`, find any matching `target.js` files, and
- * return those as entry points for the webpack output.
- *
- * A target at `./apps/main/assets/public/target.js` will create an output file
- * name `main__public.js`
- */
-
-function createEntries(entries, dir) {
-  var app = path.basename(dir);
-  var targets = glob.sync(dir + "/**/target.js");
-
-  targets.forEach(function(target) {
-    var targetName = path.basename(path.dirname(target));
-    entries[app + "__" + targetName] = [target];
-  });
-
-  return entries;
-}
+var merge = require("webpack-merge");
+var baseConfig = require('./webpack-base.config.js');
+var applicationConfig = require('./webpack-application.config.js');
 
 /**
  * Webpack configuration
  */
-module.exports = {
-
-  // Set the context as the apps directory
-  context: APPS_BASE,
-
-  // Generate the `entry` points from the filesystem
-  entry: APPS.reduce(createEntries, {}),
-
-  // Configure output
-  output: {
-    // Output into our build directory
-    path: BUILD,
-    // Template based on keys in entry above
-    // Generate hashed names for production
-    filename: "[name].js"
-  },
-
-  // Plugin/loader specific-configuration
-  plugins: [
-    new webpack.DefinePlugin({
-      DEVELOPMENT: false,
-      'process.env.NODE_ENV': '"production"'
-    }),
-    new ExtractTextPlugin("[name].css", {
-      allChunks: true
-    }),
-    new webpack.optimize.UglifyJsPlugin({
-      compress: {
-        warnings: false
-      }
-    })
-  ],
-
-  postcss: function() {
-    return {
-      defaults: [
-        cssimport({
-          addDependencyTo: webpack
-        }),
-        cssnext()
-      ]
-    };
-  },
-
-  // Quiet the output
-  stats: {
-    assets:       false,
-    assetsSort:   false,
-    cached:       false,
-    children:     false,
-    chunkModules: false,
-    chunkOrigins: false,
-    chunks:       false,
-    chunksSort:   false,
-    colors:       false,
-    errorDetails: true,
-    hash:         false,
-    modules:      false,
-    modulesSort:  false,
-    reasons:      false,
-    source:       false,
-    timings:      false,
-    version:      false
-  },
-
-  // General configuration
-  module: {
-    preLoaders: [
-      // Run all JavaScript through jshint before loading
-      {
-        test: /\.js$/,
-        exclude: /node_modules/,
-        loader: "jshint-loader"
-      }
+ module.exports = merge(
+  baseConfig({
+    quiet: true
+  }),
+  applicationConfig,
+  {
+    // Plugin/loader specific-configuration
+    plugins: [
+      new webpack.DefinePlugin({
+        DEVELOPMENT: false,
+        'process.env.NODE_ENV': '"production"'
+      }),
+      new webpack.optimize.UglifyJsPlugin({
+        compress: {
+          warnings: false
+        }
+      })
     ],
-    loaders: [
-      {
-        test: /\.(jpe?g|png|gif|svg|woff|ttf|otf|eot|ico)/,
-        loader: "file-loader?name=[path][name].[ext]"
-      },
-      {
-        test: /\.html$/,
-        loader: "html-loader"
-      },
-      {
-        test: /\.css$/,
-        // The ExtractTextPlugin pulls all CSS out into static files
-        // rather than inside the JavaScript/webpack bundle
-        loader: ExtractTextPlugin.extract("style-loader", "css-loader!postcss-loader")
-      }
-    ]
   }
-};
+);
