+++ package.json
@@ -6,37 +6,46 @@
   "repository": {},
   "dependencies": {},
   "devDependencies": {
+    "babel-core": "^6.4.0",
+    "babel-loader": "^6.2.1",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
     "css-loader": "^0.23.1",
     "del": "^1.1.1",
+    "eslint": "^1.10.3",
+    "eslint-config-standard": "^4.4.0",
+    "eslint-config-standard-react": "^1.2.1",
+    "eslint-loader": "^1.2.0",
+    "eslint-plugin-react": "^3.13.1",
+    "eslint-plugin-standard": "^1.3.1",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
-    "glob": "7.0.3",
     "gulp": "^3.8.11",
     "gulp-jsoncombine": "^1.0.2",
     "gulp-md5": "^0.1.0",
     "gulp-rev-all": "^0.7.6",
     "gulp-rev-rails-manifest": "0.0.1",
     "gulp-tap": "^0.1.3",
-    "html-loader": "^0.4.3",
     "imports-loader": "^0.6.5",
     "js-yaml": "^1.0.2",
-    "jshint": "^2.8.0",
-    "jshint-loader": "^0.8.3",
-    "postcss": "^5.0.19",
-    "postcss-cssnext": "^2.5.2",
-    "postcss-import": "^8.1.0",
-    "postcss-loader": "^0.9.1",
+    "postcss": "^5.0.14",
+    "postcss-cssnext": "^2.4.0",
+    "postcss-import": "^7.1.3",
+    "postcss-loader": "^0.8.0",
+    "postcss-modules-values": "^1.1.1",
     "run-sequence": "^1.0.2",
-    "style-loader": "^0.8.3",
+    "style-loader": "^0.13.0",
     "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1",
+    "webpack-dev-server": "1.14.0",
+    "webpack-merge": "^0.12.0",
     "webpack-notifier": "^1.2.1"
   },
   "engines": {
-    "node": ">= 4.4.2"
+    "node": ">= 4.4.2",
+    "npm": "3.8.6"
   },
   "scripts": {
-    "clean": "rm -rf ./node_modules && npm run clean-build",
+    "clean": "npm run clean-build",
     "clean-build": "rm -rf ./assets/build/*",
     "start": "node ./assets/development-server.js",
     "precompile": "node ./assets/precompile.js",
@@ -44,6 +53,7 @@
     "build-development": "webpack --config ./assets/webpack-development.config.js",
     "prebuild-production": "npm run clean-build",
     "build-production": "NODE_ENV=production webpack --config ./assets/webpack-production.config.js -p --bail",
-    "postbuild-production": "npm run precompile"
+    "postbuild-production": "npm run precompile",
+    "lint": "eslint 'apps/**/assets/**/*.js'; exit 0"
   }
 }
